Addressed code review feedback, hard coded steel thread address, moved to same package
