+++ community-care-eligibility/src/main/java/gov/va/api/health/communitycareeligibility/service/Exceptions.java
@@ -3,15 +3,15 @@ package gov.va.api.health.communitycareeligibility.service;
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
-public final class Exceptions {
-  public static final class EeUnavailableException extends RuntimeException {
-    public EeUnavailableException(Throwable cause) {
+final class Exceptions {
+  static final class EeUnavailableException extends RuntimeException {
+    EeUnavailableException(Throwable cause) {
       super("E&E is not available: " + cause.getMessage(), cause);
     }
   }
 
-  public static final class FacilitiesUnavailableException extends RuntimeException {
-    public FacilitiesUnavailableException(Throwable cause) {
+  static final class FacilitiesUnavailableException extends RuntimeException {
+    FacilitiesUnavailableException(Throwable cause) {
       super("Facilities API is not available: " + cause.getMessage(), cause);
     }
   }
