+++ community-care-eligibility/src/main/resources/application.properties
@@ -23,8 +23,3 @@ ee.truststore.password=unset
 ee.truststore.path=unset
 
 health-check.icn=skip
-health-check.address.street=46 W New Haven Ave
-health-check.address.city=Melbourne
-health-check.address.state=FL
-health-check.address.zip=32901
-health-check.drive-minutes=60
